+++ CHANGELOG.md
@@ -37,7 +37,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 * Insert
   * insert date sequence
     * DateTime `round-trip` format: https://docs.microsoft.com/en-us/dotnet/standard/base-types/standard-date-and-time-format-strings#Roundtrip
-    * improve command picker, show date format preview
     * add option to use times 12/24 formats
   * famous excerpts, e.g. The Divine Comedy, Hamlet etc...
   * excerpts in languages other than English
@@ -59,6 +58,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
   * allow to update position without reloading the entire window
 
 ## Log
+
+## [version] - 2021-04-25
+
+### Changed
+
+* Updated the `Insert Date` picker, show a sample date formatted according to the selection
+* Updated the `Insert Random` picker, show a sample string based on the selection
+* Update the `Insert Currency` picker, show a sample string based on the selection
 
 ## [0.13.0] - 2021-04-25